requirements.txt
numpy
matplotlib
scipy
pyaudiowpatch
pyserial
pyqt6